Handle boundary and multiple-parent commits
When git-blame is producing output the commits listed can be non-linear
because the development was along multiple branches.

  B -- 4 -- 3 -- 2 -- 1
        \
         9 -- 8 -- 7 -- 6

The commit "B" will be reported by git-blame as a "boundary" commit,
which we use as an indicator to not record commit 6 as a parent of
commit "B", otherwise the current commit will be added to the previous
commit's parent list.